requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
accelerate=0.31.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.14=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.29.2=pypi_0
anyio=4.8.0=pypi_0
anytree=2.12.1=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pypi_0
autograd=1.7.0=pypi_0
banks=2.1.2=pypi_0
beartype=0.18.5=pypi_0
beautifulsoup4=4.13.3=pypi_0
blis=1.2.0=pypi_0
boto3=1.37.15=pypi_0
botocore=1.37.15=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.12.14=hf0a4a13_0
catalogue=2.0.10=pypi_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpathlib=0.20.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_1
confection=0.1.5=pypi_0
contourpy=1.3.1=pypi_0
cssselect=1.3.0=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.11=pypi_0
dataclasses-json=0.6.7=pypi_0
datasets=3.5.0=pypi_0
debugpy=1.8.11=py311h313beb8_0
decorator=5.1.1=pyhd8ed1ab_1
deprecated=1.2.18=pypi_0
dill=0.3.8=pypi_0
dirtyjson=1.0.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
einops=0.8.1=pypi_0
en-core-web-sm=3.8.0=pypi_0
et-xmlfile=2.0.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
fasttext=0.9.3=pypi_0
feedfinder2=0.0.4=pypi_0
feedparser=6.0.11=pypi_0
filelock=3.17.0=pypi_0
filetype=1.2.0=pypi_0
fonttools=4.55.3=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
future=1.0.0=pypi_0
gensim=4.3.3=pypi_0
googletrans=4.0.2=pypi_0
graspologic=3.4.1=pypi_0
graspologic-native=1.2.1=pypi_0
greenlet=3.2.1=pypi_0
griffe=1.7.3=pypi_0
h11=0.14.0=pypi_0
h2=4.2.0=pypi_0
hpack=4.1.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.27.1=pypi_0
hyperframe=6.1.0=pypi_0
hyppo=0.4.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
ipykernel=6.29.5=pyh57ce528_0
ipython=8.31.0=pyh707e725_0
jedi=0.19.2=pyhd8ed1ab_1
jieba3k=0.35.1=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jpype1=1.5.1=pypi_0
json-repair=0.30.3=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.8=pypi_0
konlpy=0.6.0=pypi_0
krb5=1.21.3=h237132a_0
langchain=0.3.24=pypi_0
langchain-core=0.3.57=pypi_0
langchain-text-splitters=0.3.8=pypi_0
langcodes=3.5.0=pypi_0
langsmith=0.3.40=pypi_0
language-data=1.3.0=pypi_0
libcxx=19.1.6=ha82da77_1
libedit=3.1.20191231=hc8eb9b7_2
libffi=3.4.4=hca03da5_1
libsodium=1.0.20=h99b78c6_0
lightning-utilities=0.14.1=pypi_0
litellm=1.63.11=pypi_0
llama-cloud=0.1.19=pypi_0
llama-cloud-services=0.6.21=pypi_0
llama-index=0.12.34=pypi_0
llama-index-agent-openai=0.4.7=pypi_0
llama-index-cli=0.4.1=pypi_0
llama-index-core=0.12.34.post1=pypi_0
llama-index-embeddings-openai=0.3.1=pypi_0
llama-index-indices-managed-llama-cloud=0.6.11=pypi_0
llama-index-llms-openai=0.3.38=pypi_0
llama-index-multi-modal-llms-openai=0.4.3=pypi_0
llama-index-program-openai=0.3.1=pypi_0
llama-index-question-gen-openai=0.3.0=pypi_0
llama-index-readers-file=0.4.7=pypi_0
llama-index-readers-llama-parse=0.4.0=pypi_0
llama-parse=0.6.21=pypi_0
llvmlite=0.43.0=pypi_0
loguru=0.7.3=pypi_0
lxml=5.3.0=pypi_0
lxml-html-clean=0.4.2=pypi_0
marisa-trie=1.2.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.2.0=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.12=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h313beb8_0
neo4j=5.27.0=pypi_0
neo4j-graphrag=1.3.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
newspaper3k=0.2.8=pypi_0
nltk=3.9.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
openai=1.66.5=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.4.0=h81ee809_1
orjson=3.10.18=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
patsy=1.0.1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=24.2=py311hca03da5_0
platformdirs=4.3.8=pypi_0
plotly=5.24.1=pypi_0
pot=0.9.5=pypi_0
preshed=3.0.9=pypi_0
prompt-toolkit=3.0.48=pyha770c72_1
propcache=0.3.0=pypi_0
psutil=5.9.0=py311h80987f9_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pybind11=2.13.6=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pynndescent=0.5.13=pypi_0
pyparsing=3.2.1=pypi_0
pypdf=5.1.0=pypi_0
python=3.11.11=hb885b13_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
pytorch-lightning=2.5.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py311h313beb8_0
ragchecker=0.1.9=pypi_0
rank-bm25=0.2.2=pypi_0
readline=8.2=h1a28f6b_0
refchecker=0.2.14=pypi_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-file=2.1.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.23.1=pypi_0
s3transfer=0.11.4=pypi_0
safetensors=0.5.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
sentence-transformers=3.3.1=pypi_0
setuptools=75.1.0=py311hca03da5_0
sgmllib3k=1.0.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhd8ed1ab_0
smart-open=7.1.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
spacy=3.8.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h80987f9_0
srsly=2.5.1=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.4=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
thinc=8.3.4=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
tinysegmenter=0.3=pypi_0
tk=8.6.14=h6ba3021_0
tldextract=5.3.0=pypi_0
tokenizers=0.21.0=pypi_0
torch=2.5.1=pypi_0
torchmetrics=1.6.3=pypi_0
tornado=6.4.2=py311h80987f9_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.48.1=pypi_0
typer=0.15.1=pypi_0
types-pyyaml=6.0.12.20241230=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024.2=pypi_0
umap-learn=0.5.7=pypi_0
urllib3=2.3.0=pypi_0
wasabi=1.1.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
weasel=0.4.1=pypi_0
wheel=0.44.0=py311hca03da5_0
wikipedia=1.4.0=pypi_0
wikipedia-api=0.8.1=pypi_0
wrapt=1.17.2=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.18.3=pypi_0
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h18a0788_1
zstandard=0.23.0=pypi_0
